+++ mtpgsql/src/backend/utils/adt/varbit.c
@@ -115,8 +115,8 @@ zpbit_in(char *s, int dummy, int32 atttypmod)
 	len = VARBITDATALEN(atttypmod);
 
 	if (len > MaxAttrSize)
-		elog(ERROR, "zpbit_in: length of bit() must be less than %ld",
-			 (MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE);
+		elog(ERROR, "zpbit_in: length of bit() must be less than %u",
+             (uint32)(MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE);
 
 	result = (bits8 *) palloc(len);
 	/* set to 0 so that *r is always initialised and strin is zero-padded */
@@ -346,8 +346,8 @@ varbit_in(char *s, int dummy, int32 atttypmod)
 	len = VARBITDATALEN(bitlen);
 
 	if (len > MaxAttrSize)
-		elog(ERROR, "varbit_in: length of bit() must be less than %ld",
-			 (MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE);
+		elog(ERROR, "varbit_in: length of bit() must be less than %u",
+             (uint32)(MaxAttrSize - VARHDRSZ - VARBITHDRSZ) * BITSPERBYTE);
 
 	result = (bits8 *) palloc(len);
 	/* set to 0 so that *r is always initialised and strin is zero-padded */
